Organize helper functions better
